requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
